+++ CHANGELOG.md
@@ -80,4 +80,5 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 - Echo Glass now emits the `minecraft:block_change` Game Event upon cracking.
 - Likely fixed an issue that caused Jellyfish to be invisible with certain Shader Packs.
   - This changes the `RenderType` of Jellyfish to use a Vanilla one. As a result, the visuals of Jellyfish may be different than intended.
+- Fixed crashes that could occur when trying to render a Wilder Wild mob that has variants in a context where dynamic registries aren't available. ([#558](https://github.com/FrozenBlock/WilderWild/issues/558))
 - Migrated the implementation of animated Sculk Sensor Block Entities from a mixin to Fabric's and NeoForge's Data Attachments.